facile

Attribute Macro facade

#[facade]

Construct a façade trait based on an implementation.

Given an implementation of a non-existent trait, extract a trait by that name from the signature of the implementation, and implement it as described.

Parameters

visibility

The visibility of the resulting trait, e.g. pub or pub(super). If no visibility is provided then no explicit visibility will be applied to the trait.

where

A list of where predicates to apply to the trait definition, e.g. T: Clone, Self: Send. Generics on the implementation (impl<…> where …) apply to the implementation rather than the trait. If you want to constrain the trait itself, use this parameter. Bounds on Self can be used to provide supertraits (which are otherwise not supported by the syntax).

Panics

If the input is not a trait implementation.

Type parameters and associated types

Type or lifetime parameters on the trait will be automatically lifted to the trait definition, and can be constrained with where.

Since type parameters on the type aren’t in scope for the trait, you must assign them to an associated type and reference them through that (as Self::AssocType).

In an extension of the usual impl syntax, associated types can be given both bounds (for the trait definitiion) and definitions (for the implementation), e.g.

type Future: Future = std::future::Ready<()>;

Default implementations

The facade(default) attribute can be used to indicate that a function depends only on the trait definition itself rather than any detail of the implementation, and should be provided as a default implementation on the trait. This can be useful if you’re planning to add more implementations of the trait.

Example

Put the facile::facade annotation on a trait implementation with a non-existent trait name to generate a trait by that name and also implement it as written.

struct FooImpl<F, P> {
    future: F,
    pointer: P,
}

#[facile::facade(
    visibility = pub(crate),
    where(F: Future<Output = ()>),
)]
impl<F, P> Foo<F> for FooImpl<F, P>
where
    F: Future<Output = ()> + Clone,
    P: AsRef<str>,
{
    type Future: Future<Output = ()> = F;

    fn future(&self) -> Self::Future {
        self.future.clone()
    }

    fn as_str(&self) -> &str {
        self.pointer.as_ref()
    }

    #[facade(default)]
    async fn run(&self) {
        println!("running future: {}", self.as_str());
        self.future().await
    }
}

struct OtherFoo;
impl Foo<std::future::Ready<()>> for OtherFoo {
    type Future = std::future::Ready<()>;
    fn future(&self) -> Self::Future {
        std::future::ready(())
    }
    fn as_str(&self) -> &str {
        "OtherFoo"
    }
}

async fn use_foo(foo: impl Foo<std::future::Ready<()>>) {
    foo.run().await
}